rename tests/ dir to functional_tests
