+++ functional_tests/databasecockroachdb_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestDatabaseCockroachDB(t *testing.T) {
+	// TODO: Add functional_tests for DatabaseCockroachDB method
+}
