+++ functional_tests/databasemariadb_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestDatabaseMariaDB(t *testing.T) {
+	// TODO: Add functional_tests for DatabaseMariaDB method
+}
