+++ functional_tests/databasemssql_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestDatabaseMSSQL(t *testing.T) {
+	// TODO: Add functional_tests for DatabaseMSSQL method
+}
