+++ functional_tests/databasemysql_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestDatabaseMySQL(t *testing.T) {
+	// TODO: Add functional_tests for DatabaseMySQL method
+}
