+++ functional_tests/databaseoracle_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestDatabaseOracle(t *testing.T) {
+	// TODO: Add functional_tests for DatabaseOracle method
+}
