+++ functional_tests/databasepostgres_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestDatabasePostgres(t *testing.T) {
+	// TODO: Add functional_tests for DatabasePostgres method
+}
