+++ functional_tests/databasesqllite_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestDatabaseSqlLite(t *testing.T) {
+	// TODO: Add functional_tests for DatabaseSqlLite method
+}
