+++ functional_tests/fetch_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestFetch(t *testing.T) {
+	// TODO: Add functional_tests for Fetch method
+}
