+++ functional_tests/hydrate_test.go
@@ -1,0 +1,7 @@
+package functional_tests
+
+import "testing"
+
+func TestHydrate(t *testing.T) {
+	// TODO: Add functional_tests for Hydrate method
+}
