--- tests/databasecockroachdb_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestDatabaseCockroachDB(t *testing.T) {
-	// TODO: Add tests for DatabaseCockroachDB method
-}
