--- tests/databasemariadb_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestDatabaseMariaDB(t *testing.T) {
-	// TODO: Add tests for DatabaseMariaDB method
-}
