--- tests/databasemssql_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestDatabaseMSSQL(t *testing.T) {
-	// TODO: Add tests for DatabaseMSSQL method
-}
