--- tests/databasemysql_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestDatabaseMySQL(t *testing.T) {
-	// TODO: Add tests for DatabaseMySQL method
-}
