--- tests/databaseoracle_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestDatabaseOracle(t *testing.T) {
-	// TODO: Add tests for DatabaseOracle method
-}
