--- tests/databasepostgres_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestDatabasePostgres(t *testing.T) {
-	// TODO: Add tests for DatabasePostgres method
-}
