--- tests/databasesqllite_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestDatabaseSqlLite(t *testing.T) {
-	// TODO: Add tests for DatabaseSqlLite method
-}
