--- tests/fetch_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestFetch(t *testing.T) {
-	// TODO: Add tests for Fetch method
-}
