--- tests/hydrate_test.go
@@ -1,7 +1,0 @@
-package tests
-
-import "testing"
-
-func TestHydrate(t *testing.T) {
-	// TODO: Add tests for Hydrate method
-}
